fix: Move bot config into seperate module
