+++ src/config.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 import argparse
 import os
 from dataclasses import dataclass
